Add custom Info.plist to fix issue #142 "Logisim Evolution forces the dGPU to be used"
The custom Info.plist differs from the auto-generated Info.plist only one point. The following key is added:
```
  <key>NSSupportsAutomaticGraphicsSwitching</key>
  <string>true</string>
```
There does not seem to exist yet a mechanism to inject such a small setting through a custom setting in ```build.xml```.

Be aware, this file needs to be updated when changing the version number of Logisim-evolution to stay in sync.