Remove lablib-checkboxtree-3.3.jar
Unfortunately, `lablib-checkboxtree-3.3.jar` is not available on
Maven Central, but a 3rd-party repository on GitHub was found.
There's a certain risk that this one will be deleted by its maintainer.
We could consider forking it to just maintain this last external
and outdated JAR dependency until it can be replaced with a suitable
alternative.