Bug in memry Poke
Fixed a bug in the key action when editing memories with the Poke tool, which interfered with the canvas.
Standardises the use of keyboard keys in the two hexadecimal editors of the memory contents.
According to these rules :
 The Down Arrow or Enter key will scroll down one line.
The Up Arrow or Ctrl+Enter key will scroll up one line.
The mouse wheel is used for the same scrolling functions.
The Right Arrow or Space keys will scroll one box to the right.
The Left Arrow or Backspace keys will scroll one box to the left.
The Page Down or Ctrl+Space keys will scroll down one page.
The Page Up or Ctrl+Return keys will scroll up one page.
The Del resets the selected cell with a zero