Improve enable attribute for multiplexer
Instead of adding a new INVERT_ENABLE to the multiplexer and PlexersLibrary, we add a new attribute to the PlexersLibrary named ENABLE_TYPE. This attribute is intended to replace the enable attribute for plexer components that support inverted enable. Thus, the enable type attribute clashes in ID with the  enable attribute; they are not intended to be used together.