translations: fix wrong zh-cn names of the TTL series
There are some obvious typos or incorrect Chinese translations in
the TTL series. My work includes standardizing the common descriptions
of the inputs and outputs, and fixing the incorrect naming, such as
the octal TTL74541 buffers, which was translated into something that
it only works from 0 to 7. I rename the octal to 8-bit in Chinese.