Add Real-Time Clock component
Introduces a new Real-Time Clock component to the IO library, including its implementation, icon, and localized strings. The component outputs the current time in either epoch milliseconds or seconds, configurable via attributes.

The component automatically updates it's value without the need for a clock signal. This can be changed in future commits if it creates any problems