Fix multiplier partial result
The partial result of multiplication was using integers was using ints and was not being calculated correctly.

The calculations were simplified overall and the "extend" method was moved to the Value class as toBigInteger(), and then optimized. This method was moved to Value because it will be used by other arithmetic components and it makes more sense having it along with other value conversion methods